requirements.txt
click
twitchio
requests
python-dotenv
jmespath
pymongo
